+++ saleor/product/filters.py
@@ -1,3 +1,5 @@
+import functools
+import operator
 from collections import OrderedDict
 
 from django.db.models import Q
@@ -15,6 +17,138 @@ SORT_BY_FIELDS = OrderedDict([
         'Product list sorting option', 'last updated'))])
 
 
+def serialize_attribute(value):
+    """Serialize attribute as a tuple to a string."""
+    if isinstance(value, tuple) and len(value) == 2:
+        return ':'.join(value)
+    return None
+
+
+def parse_attribute(value):
+    """Parse attribute as a string to a tuple."""
+    if isinstance(value, str):
+        split = value.split(':')
+        if len(split) == 2:
+            return tuple(split)
+    return None
+
+
+class MergedAttributes:
+    """Merge attribute queryset into a dictionary by attribute slug.
+    Note that the attribute values should be prefetched for
+    optimizing performance."""
+    def __init__(self, attributes):
+        self._merged_attributes = self._get_merged_attributes(attributes)
+
+    def _get_merged_attributes(self, attributes):
+        """Merge attributes with same slug into one item in dictionary."""
+        merged_attributes = {}
+        for attribute in attributes:
+            same_slug_attrs = merged_attributes.setdefault(attribute.slug, [])
+            same_slug_attrs.append(attribute)
+        return merged_attributes
+
+    def get_kvs(self, attr_slug, value_slug):
+        """Get a list of attribute-pk/value-pk pairs
+        for attributes with same attribute slug and value slug."""
+        kvs = []
+        for attr in self._merged_attributes[attr_slug]:
+            for value in attr.values.all():
+                if value_slug == value.slug:
+                    kvs.append((attr.pk, value.pk))
+        return kvs
+
+    def get_attributes(self):
+        """Get attributes as a dict."""
+        return self._merged_attributes
+
+    def get_values(self, attr_slug):
+        """Get attributes values for attributes with same attribute slug."""
+        values = dict()
+        for attr in self._merged_attributes[attr_slug]:
+            for value in attr.values.all():
+                values.setdefault(value.slug, []).append(value)
+        return values
+
+    def get_choices(self, attr_slug):
+        """Get attributes values as choices for MultipleChoiceFilter."""
+        values, choices = self.get_values(attr_slug), []
+        for value_slug, value_list in values.items():
+            v = serialize_attribute((attr_slug, value_slug))
+            # By default the translated name of the first one in
+            # attribute values with same slug is used
+            label = value_list[0].translated.name
+            choices.append((v, label))
+        return choices
+
+    def _product_attr_or_product_variant_attr_query(self, attr_k, attr_v):
+        product_attr_query = Q(**{'attributes__%s' % (attr_k, ): attr_v})
+        variant_attr_query = Q(
+            **{'variants__attributes__%s' % (attr_k, ): attr_v})
+        return product_attr_query | variant_attr_query
+
+    def get_query(self, attr_slug, value_slug):
+        """Get query for attributes with same attribute slug and value slug."""
+        # Validate attribute slug and value slug
+        if attr_slug not in self.get_attributes():
+            raise ValueError('Unknown attribute slug: %r' % (attr_slug,))
+        if value_slug not in self.get_values(attr_slug):
+            raise ValueError(
+                'Unknown attribute value slug: %r' % (value_slug,))
+
+        # Combine filters of attributes
+        # with same attribute slug and value slug using OR operator
+        kvs = self.get_kvs(attr_slug, value_slug)
+        query = functools.reduce(
+            operator.or_,
+            [self._product_attr_or_product_variant_attr_query(
+                attr_pk, value_pk) for attr_pk, value_pk in kvs])
+
+        return query
+
+
+class AttributeMultipleChoiceFilter(MultipleChoiceFilter):
+    def __init__(self, merged_attributes, *args, **kwargs):
+        self._merged_attributes = merged_attributes
+        super().__init__(*args, **kwargs)
+
+    def filter(self, qs, value):
+        """Copied from django-filter directly but only modified
+        get_filter_predicate which supports returning query directly."""
+        if not value:
+            # Even though not a noop, no point filtering if empty.
+            return qs
+
+        if self.is_noop(qs, value):
+            return qs
+
+        if not self.conjoined:
+            q = Q()
+        for v in set(value):
+            if v == self.null_value:
+                v = None
+            predicate = self.get_filter_predicate(v)
+            if self.conjoined:
+                qs = self.get_method(qs)(predicate)
+            else:
+                q |= Q(predicate)
+
+        if not self.conjoined:
+            qs = self.get_method(qs)(q)
+
+        return qs.distinct() if self.distinct else qs
+
+    def get_filter_predicate(self, v):
+        # Parse attribute scalar
+        attribute = parse_attribute(v)
+        if attribute is None:
+            raise ValueError(
+                'Unknown attribute scalar: %r' % (attribute,))
+
+        attr_slug, value_slug = attribute
+        return self._merged_attributes.get_query(attr_slug, value_slug)
+
+
 class ProductFilter(SortedFilterSet):
     sort_by = OrderingFilter(
         label=pgettext_lazy('Product list sorting form', 'Sort by'),
@@ -29,26 +163,21 @@ class ProductFilter(SortedFilterSet):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.product_attributes, self.variant_attributes = (
-            self._get_attributes())
-        self.filters.update(self._get_product_attributes_filters())
-        self.filters.update(self._get_product_variants_attributes_filters())
+        self._merged_attributes = self._get_merged_attributes()
+        self.filters.update(self._get_attributes_filters())
         self.filters = OrderedDict(sorted(self.filters.items()))
 
-    def _get_attributes(self):
+    def _get_merged_attributes(self):
         q_product_attributes = self._get_product_attributes_lookup()
         q_variant_attributes = self._get_variant_attributes_lookup()
-        product_attributes = (
+        attributes = (
             Attribute.objects.all()
             .prefetch_related('translations', 'values__translations')
-            .filter(q_product_attributes)
+            .filter(q_product_attributes | q_variant_attributes)
             .distinct())
-        variant_attributes = (
-            Attribute.objects.all()
-            .prefetch_related('translations', 'values__translations')
-            .filter(q_variant_attributes)
-            .distinct())
-        return product_attributes, variant_attributes
+
+        merged_attributes = MergedAttributes(attributes)
+        return merged_attributes
 
     def _get_product_attributes_lookup(self):
         raise NotImplementedError()
@@ -56,30 +185,17 @@ class ProductFilter(SortedFilterSet):
     def _get_variant_attributes_lookup(self):
         raise NotImplementedError()
 
-    def _get_product_attributes_filters(self):
-        filters = {}
-        for attribute in self.product_attributes:
-            filters[attribute.slug] = MultipleChoiceFilter(
-                field_name='attributes__%s' % attribute.pk,
-                label=attribute.translated.name,
+    def _get_attributes_filters(self):
+        filters, attributes = {}, self._merged_attributes.get_attributes()
+        for attr_slug in attributes:
+            filters[attr_slug] = AttributeMultipleChoiceFilter(
+                merged_attributes=self._merged_attributes,
+                # By default the translated name of the first one in
+                # # attributes with same slug is used
+                label=attributes[attr_slug][0].translated.name,
                 widget=CheckboxSelectMultiple,
-                choices=self._get_attribute_choices(attribute))
+                choices=self._merged_attributes.get_choices(attr_slug))
         return filters
-
-    def _get_product_variants_attributes_filters(self):
-        filters = {}
-        for attribute in self.variant_attributes:
-            filters[attribute.slug] = MultipleChoiceFilter(
-                field_name='variants__attributes__%s' % attribute.pk,
-                label=attribute.translated.name,
-                widget=CheckboxSelectMultiple,
-                choices=self._get_attribute_choices(attribute))
-        return filters
-
-    def _get_attribute_choices(self, attribute):
-        return [
-            (choice.pk, choice.translated.name)
-            for choice in attribute.values.all()]
 
 
 class ProductCategoryFilter(ProductFilter):